backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
openai>=2.21.0,<3
pdfplumber==0.10.0
python-docx==1.1.0
pillow==10.2.0
pillow-heif==0.16.0
pytesseract==0.3.10
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.26.0
structlog==24.1.0
tenacity==8.2.3
numpy==1.26.3
tiktoken==0.5.2
tinker==0.12.0
pytest==7.4.4
pytest-asyncio==0.23.4
